twitter_consumer/requirements.txt
#
####### example-requirements.txt #######
#
###### Requirements without Version Specifiers ######
pymongo
kafka-python